+++ shaders/animupdate_normals.comp.glsl
@@ -40,11 +40,6 @@ layout(push_constant) uniform animationConstantsPush
   AnimationConstants constants;
 };
 
-layout(buffer_reference, scalar) readonly buffer RenderInstances_in
-{
-  RenderInstance instances[];
-};
-
 layout(buffer_reference, scalar) readonly buffer U32Buffer
 {
   uint32_t i[];
@@ -60,7 +55,7 @@ void main()
 {
   uint32_t index = gl_GlobalInvocationID.x;
 
-  RenderInstance instance = RenderInstances_in(constants.renderInstances).instances[constants.instanceIndex];
+  RenderInstance instance = RenderInstances_in(constants.renderInstances).d[constants.instanceIndex];
 
   if(index >= instance.numTriangles)
   {