+++ shaders/cluster_blas_instances.comp.glsl
@@ -30,15 +30,24 @@
 
 #include "shaderio.h"
 
-layout(local_size_x = STATISTICS_WORKGROUP_SIZE) in;
+layout(local_size_x = CLUSTER_BLAS_WORKGROUP_SIZE) in;
 layout(push_constant, scalar) uniform pushConstant {
-  StatisticsConstants push;
+  ClusterBlasConstants push;
 };
 
 void main()
 {
-  if (gl_GlobalInvocationID.x < push.count)
+  uint idx = gl_GlobalInvocationID.x;
+
+  // update ray tracing blas address prior tlas update
+  if (idx < push.instanceCount)
   {
-    atomicAdd(push.sum.d[0], uint64_t(push.sizes.d[gl_GlobalInvocationID.x]));
-  };
+    push.rayInstances.d[idx].accelerationStructureReference = push.blasAddresses.d[ push.animated != 0 ? idx : push.instances.d[idx].geometryID ];
+  }
+
+  // for statistics we sum blas/cluster sizes
+  if (idx < push.sumCount)
+  {
+    atomicAdd(push.sum.d[0], uint64_t(push.sizes.d[idx]));
+  }
 }